ml_compare_r/requirements.txt
django==4.2
python-docx
gensim
numpy
pandas
scipy==1.12
scikit-learn
